requirements.txt
pandas
scipy
numpy
matplotlib
plotly
sklearn
streamlit-aggrid
#streamlit_book
git+https://github.com/sebastiandres/streamlit_book.git